+++ Fossil/test/TestHeader.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../src/FileIO.h"
+#include "../src/Settings.h"
+
+//	Functions to test functions in FileIO.h
+void test_logMessage();
+void test_changeSetting();
+void test_getSettings();